Correction tailwind + securité

+++ src/Controller/ClothingItemController.php
@@ -5,9 +5,12 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Entity\Partner;
 use App\Entity\ClothingItem;
+use App\Entity\AINotification;
+use App\Entity\Wardrobe;
 use App\Form\ClothingItemTypeImagePath;
 use App\Form\PartnerType;
 use App\Repository\PartnerRepository;
+use App\Repository\ClothingItemRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -39,11 +42,19 @@ final class ClothingItemController extends AbstractController{
         $brand = $partnerRepository->findOneBy(['name' => $data['brand']]);
     
         if (!$brand) {
+
+            $notif = new AINotification();
+            $notif->setCreatedAt(new \DateTimeImmutable());
+            $notif->setMessage("La marque " . $data['brand'] . " a été ajoutée dans la BDD");
+            $notif->setStatus("Pending");
+
+            $entityManager->persist($notif);
+
             $brand = new Partner();
             $formBrand = $this->createForm(PartnerType::class, $brand, [
                 'csrf_protection' => false, 
             ]);
-    
+            
             $formBrand->submit(['name' => $data['brand']], false);
     
             if ($formBrand->isSubmitted() && $formBrand->isValid()) {
@@ -78,12 +89,36 @@ final class ClothingItemController extends AbstractController{
 
     #[Route('/{id}', name: 'app_clothing_item_delete', methods: ['POST'])]
     public function delete(Request $request, ClothingItem $clothingItem, EntityManagerInterface $entityManager): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$clothingItem->getId(), $request->getPayload()->getString('_token'))) {
+    {   
+        $user = $this->getUser(); 
+    
+        if (!$user instanceof User) {
+            $this->addFlash('error', 'L\'utilisateur n\'est pas valide.');
+            return $this->redirectToRoute('home');
+        }
+    
+        $userWardrobe = $user->getWardrobe(); 
+        $itemWardrobes = $clothingItem->getWardrobes();
+    
+        $hasAccess = false;
+        foreach ($itemWardrobes as $wardrobe) {
+            if ($wardrobe->getId() === $userWardrobe->getId()) {
+                $hasAccess = true;
+                break;
+            }
+        }
+    
+        if (!$hasAccess) {
+            $this->addFlash('error', 'Permission Invalide');
+            return $this->redirectToRoute('home');
+        }
+    
+        if ($this->isCsrfTokenValid('delete' . $clothingItem->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($clothingItem);
             $entityManager->flush();
         }
-
+    
         return $this->redirectToRoute('wardrobe_list', [], Response::HTTP_SEE_OTHER);
     }
+    
 }
